Move a bunch of code out host, reorg modules slightly.
