+++ soroban-env-host/src/host/conversion.rs
@@ -1,7 +1,8 @@
 use std::rc::Rc;
 
-use super::metered_clone::MeteredClone;
-use crate::host_object::HostVec;
+use super::metered_clone::{self, MeteredClone};
+use crate::budget::AsBudget;
+use crate::host_object::{HostMap, HostObject, HostVec};
 use crate::xdr::{
     Hash, LedgerKey, LedgerKeyContractData, ScHostFnErrorCode, ScHostObjErrorCode,
     ScHostValErrorCode, ScVal, ScVec, Uint256,
@@ -9,8 +10,17 @@ use crate::xdr::{
 use crate::{events::DebugError, xdr::ContractCostType, Host, HostError, RawVal};
 use ed25519_dalek::{PublicKey, Signature, SIGNATURE_LENGTH};
 use sha2::{Digest, Sha256};
-use soroban_env_common::xdr::{self, AccountId, ScBytes};
-use soroban_env_common::{BytesObject, TryFromVal, U32Val, VecObject};
+use soroban_env_common::num::{
+    i256_from_pieces, i256_into_pieces, u256_from_pieces, u256_into_pieces,
+};
+use soroban_env_common::xdr::{
+    self, int128_helpers, AccountId, Int128Parts, Int256Parts, ScBytes, ScMap, ScMapEntry,
+    UInt128Parts, UInt256Parts,
+};
+use soroban_env_common::{
+    BytesObject, Convert, Object, ScValObjRef, ScValObject, TryFromVal, TryIntoVal, U32Val,
+    VecObject,
+};
 
 impl Host {
     // Notes on metering: free
@@ -316,5 +326,224 @@ impl Host {
 
     pub(crate) fn scbytes_from_hash(&self, hash: &Hash) -> Result<ScBytes, HostError> {
         self.scbytes_from_slice(hash.as_slice())
+    }
+}
+
+impl Convert<&Object, ScValObject> for Host {
+    type Error = HostError;
+    fn convert(&self, ob: &Object) -> Result<ScValObject, Self::Error> {
+        self.from_host_obj(*ob)
+    }
+}
+
+impl Convert<Object, ScValObject> for Host {
+    type Error = HostError;
+    fn convert(&self, ob: Object) -> Result<ScValObject, Self::Error> {
+        self.from_host_obj(ob)
+    }
+}
+
+impl<'a> Convert<&ScValObjRef<'a>, Object> for Host {
+    type Error = HostError;
+    fn convert(&self, ob: &ScValObjRef<'a>) -> Result<Object, Self::Error> {
+        self.to_host_obj(ob)
+    }
+}
+
+impl<'a> Convert<ScValObjRef<'a>, Object> for Host {
+    type Error = HostError;
+    fn convert(&self, ob: ScValObjRef<'a>) -> Result<Object, Self::Error> {
+        self.to_host_obj(&ob)
+    }
+}
+
+impl Host {
+    pub(crate) fn from_host_val(&self, val: RawVal) -> Result<ScVal, HostError> {
+        // Charges a single unit to for the RawVal -> ScVal conversion.
+        // The actual conversion logic occurs in the `common` crate, which
+        // translates a u64 into another form defined by the xdr.
+        // For an `Object`, the actual structural conversion (such as byte
+        // cloning) occurs in `from_host_obj` and is metered there.
+        self.charge_budget(ContractCostType::ValXdrConv, None)?;
+        ScVal::try_from_val(self, &val)
+            .map_err(|_| self.err_status(ScHostValErrorCode::UnknownError))
+    }
+
+    pub(crate) fn to_host_val(&self, v: &ScVal) -> Result<RawVal, HostError> {
+        // `ValXdrConv` is const cost in both cpu and mem. The input=0 will be ignored.
+        self.charge_budget(ContractCostType::ValXdrConv, None)?;
+        v.try_into_val(self)
+            .map_err(|_| self.err_status(ScHostValErrorCode::UnknownError))
+    }
+
+    pub(crate) fn from_host_obj(&self, ob: impl Into<Object>) -> Result<ScValObject, HostError> {
+        unsafe {
+            self.unchecked_visit_val_obj(ob.into(), |ob| {
+                // This accounts for conversion of "primitive" objects (e.g U64)
+                // and the "shell" of a complex object (ScMap). Any non-trivial
+                // work such as byte cloning, has to be accounted for and
+                // metered in indivial match arms.
+                // `ValXdrConv` is const cost in both cpu and mem. The input=0 will be ignored.
+                self.charge_budget(ContractCostType::ValXdrConv, None)?;
+                let val = match ob {
+                    None => {
+                        return Err(self.err_status(ScHostObjErrorCode::UnknownReference));
+                    }
+                    Some(ho) => match ho {
+                        HostObject::Vec(vv) => {
+                            metered_clone::charge_heap_alloc::<ScVal>(
+                                vv.len() as u64,
+                                self.as_budget(),
+                            )?;
+                            let sv = vv.iter().map(|e| self.from_host_val(*e)).collect::<Result<
+                                Vec<ScVal>,
+                                HostError,
+                            >>(
+                            )?;
+                            ScVal::Vec(Some(ScVec(self.map_err(sv.try_into())?)))
+                        }
+                        HostObject::Map(mm) => {
+                            metered_clone::charge_heap_alloc::<ScMapEntry>(
+                                mm.len() as u64,
+                                self.as_budget(),
+                            )?;
+                            let mut mv = Vec::with_capacity(mm.len());
+                            for (k, v) in mm.iter(self)? {
+                                let key = self.from_host_val(*k)?;
+                                let val = self.from_host_val(*v)?;
+                                mv.push(ScMapEntry { key, val });
+                            }
+                            ScVal::Map(Some(ScMap(self.map_err(mv.try_into())?)))
+                        }
+                        HostObject::U64(u) => ScVal::U64(*u),
+                        HostObject::I64(i) => ScVal::I64(*i),
+                        HostObject::TimePoint(tp) => {
+                            ScVal::Timepoint(tp.metered_clone(self.as_budget())?)
+                        }
+                        HostObject::Duration(d) => {
+                            ScVal::Duration(d.metered_clone(self.as_budget())?)
+                        }
+                        HostObject::U128(u) => ScVal::U128(UInt128Parts {
+                            hi: int128_helpers::u128_hi(*u),
+                            lo: int128_helpers::u128_lo(*u),
+                        }),
+                        HostObject::I128(i) => ScVal::I128(Int128Parts {
+                            hi: int128_helpers::i128_hi(*i),
+                            lo: int128_helpers::i128_lo(*i),
+                        }),
+                        HostObject::U256(u) => {
+                            let (hi_hi, hi_lo, lo_hi, lo_lo) = u256_into_pieces(*u);
+                            ScVal::U256(UInt256Parts {
+                                hi_hi,
+                                hi_lo,
+                                lo_hi,
+                                lo_lo,
+                            })
+                        }
+                        HostObject::I256(i) => {
+                            let (hi_hi, hi_lo, lo_hi, lo_lo) = i256_into_pieces(*i);
+                            ScVal::I256(Int256Parts {
+                                hi_hi,
+                                hi_lo,
+                                lo_hi,
+                                lo_lo,
+                            })
+                        }
+                        HostObject::Bytes(b) => ScVal::Bytes(b.metered_clone(self.as_budget())?),
+                        HostObject::String(s) => ScVal::String(s.metered_clone(self.as_budget())?),
+                        HostObject::Symbol(s) => ScVal::Symbol(s.metered_clone(self.as_budget())?),
+                        HostObject::ContractExecutable(cc) => {
+                            ScVal::ContractExecutable(cc.metered_clone(self.as_budget())?)
+                        }
+                        HostObject::Address(addr) => {
+                            ScVal::Address(addr.metered_clone(self.as_budget())?)
+                        }
+                        HostObject::NonceKey(nk) => {
+                            ScVal::LedgerKeyNonce(nk.metered_clone(self.as_budget())?)
+                        }
+                    },
+                };
+                Ok(ScValObject::unchecked_from_val(val))
+            })
+        }
+    }
+
+    pub(crate) fn to_host_obj(&self, ob: &ScValObjRef<'_>) -> Result<Object, HostError> {
+        // `ValXdrConv` is const cost in both cpu and mem. The input=0 will be ignored.
+        self.charge_budget(ContractCostType::ValXdrConv, None)?;
+        let val: &ScVal = (*ob).into();
+        match val {
+            ScVal::Vec(Some(v)) => {
+                metered_clone::charge_heap_alloc::<RawVal>(v.len() as u64, self.as_budget())?;
+                let mut vv = Vec::with_capacity(v.len());
+                for e in v.iter() {
+                    vv.push(self.to_host_val(e)?)
+                }
+                Ok(self.add_host_object(HostVec::from_vec(vv)?)?.into())
+            }
+            ScVal::Map(Some(m)) => {
+                metered_clone::charge_heap_alloc::<(RawVal, RawVal)>(
+                    m.len() as u64,
+                    self.as_budget(),
+                )?;
+                let mut mm = Vec::with_capacity(m.len());
+                for pair in m.iter() {
+                    let k = self.to_host_val(&pair.key)?;
+                    let v = self.to_host_val(&pair.val)?;
+                    mm.push((k, v))
+                }
+                Ok(self.add_host_object(HostMap::from_map(mm, self)?)?.into())
+            }
+            ScVal::Vec(None) | ScVal::Map(None) => {
+                Err(self.err_status(ScHostValErrorCode::MissingObject))
+            }
+            ScVal::U64(u) => Ok(self.add_host_object(*u)?.into()),
+            ScVal::I64(i) => Ok(self.add_host_object(*i)?.into()),
+            ScVal::Timepoint(t) => Ok(self
+                .add_host_object(t.metered_clone(self.as_budget())?)?
+                .into()),
+            ScVal::Duration(d) => Ok(self
+                .add_host_object(d.metered_clone(self.as_budget())?)?
+                .into()),
+            ScVal::U128(u) => Ok(self
+                .add_host_object(int128_helpers::u128_from_pieces(u.hi, u.lo))?
+                .into()),
+            ScVal::I128(i) => Ok(self
+                .add_host_object(int128_helpers::i128_from_pieces(i.hi, i.lo))?
+                .into()),
+            ScVal::U256(u) => Ok(self
+                .add_host_object(u256_from_pieces(u.hi_hi, u.hi_lo, u.lo_hi, u.lo_lo))?
+                .into()),
+            ScVal::I256(i) => Ok(self
+                .add_host_object(i256_from_pieces(i.hi_hi, i.hi_lo, i.lo_hi, i.lo_lo))?
+                .into()),
+            ScVal::Bytes(b) => Ok(self
+                .add_host_object(b.metered_clone(self.as_budget())?)?
+                .into()),
+            ScVal::String(s) => Ok(self
+                .add_host_object(s.metered_clone(self.as_budget())?)?
+                .into()),
+            ScVal::Symbol(s) => Ok(self
+                .add_host_object(s.metered_clone(self.as_budget())?)?
+                .into()),
+            ScVal::ContractExecutable(cc) => Ok(self
+                .add_host_object(cc.metered_clone(self.as_budget())?)?
+                .into()),
+            ScVal::LedgerKeyNonce(_) => {
+                Err(self.err_general("nonce keys aren't allowed to be used directly"))
+            }
+            ScVal::Address(addr) => Ok(self
+                .add_host_object(addr.metered_clone(self.as_budget())?)?
+                .into()),
+
+            ScVal::Bool(_)
+            | ScVal::Void
+            | ScVal::Status(_)
+            | ScVal::U32(_)
+            | ScVal::I32(_)
+            | ScVal::LedgerKeyContractExecutable => {
+                Err(self.err_status(ScHostObjErrorCode::UnexpectedType))
+            }
+        }
     }
 }
